promote tag management commands as system tools removed unittest for tags
